+++ packages/eslint-plugin/index.d.ts
@@ -1,4 +1,4 @@
-import { TSESLint } from '@typescript-eslint/experimental-utils';
+import { TSESLint } from '@typescript-eslint/utils';
 
 export const rules: Record<string, TSESLint.RuleModule<string, unknown[]>>;
 export const configs: Record<string, TSESLint.Linter.Config>;