+++ packages/eslint-plugin/src/rules/no-magic-numbers.ts
@@ -1,7 +1,4 @@
-import {
-  AST_NODE_TYPES,
-  TSESTree,
-} from '@typescript-eslint/experimental-utils';
+import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/utils';
 import * as util from '../util';
 import { getESLintCoreRule } from '../util/getESLintCoreRule';
 
@@ -12,7 +9,10 @@ type MessageIds = util.InferMessageIdsTypeFromRule<typeof baseRule>;
 
 // Extend base schema with additional property to ignore TS numeric literal types
 const schema = util.deepMerge(
-  { ...baseRule.meta.schema },
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- https://github.com/microsoft/TypeScript/issues/17002
+  Array.isArray(baseRule.meta.schema)
+    ? baseRule.meta.schema[0]
+    : baseRule.meta.schema,
   {
     properties: {
       ignoreNumericLiteralTypes: {
@@ -72,7 +72,7 @@ export default util.createRule<Options, MessageIds>({
 
         // Check if the node is a readonly class property
         if (
-          typeof node.value === 'number' &&
+          (typeof node.value === 'number' || typeof node.value === 'bigint') &&
           isParentTSReadonlyPropertyDefinition(node)
         ) {
           if (options.ignoreReadonlyClassProperties) {