+++ packages/eslint-plugin/tests/rules/no-unnecessary-qualifier.test.ts
@@ -1,4 +1,4 @@
-import { AST_NODE_TYPES } from '@typescript-eslint/experimental-utils';
+import { AST_NODE_TYPES } from '@typescript-eslint/utils';
 import rule from '../../src/rules/no-unnecessary-qualifier';
 import { RuleTester, getFixturesRootDir } from '../RuleTester';